Add npm script to rebuild lerna packages when changes are detected
Remove the need to manually run `npm run bootstrap` when working
on code in the `packages/` folders.

Add a new `start:watch` npm script which runs the following in parallel:
- `npm start` (existing script to run the webpack dev server)
- `npm run bootstrap:watch`
  this is a new script that watches the src folders of the lerna packages
  for change and runs `npm run bootstrap`. The webpack dev server will
  detect the changes and recompile as needed.

This is a first pass to improve the dev workflow when modifying
packages/dashboard-components or dashboard-utils and testing those
changes in the main dashboard project. This is in addition to the
hot reloading provided in the storybook for the components package.

A future update may change the approach here to avoid the additional
filesystem watcher and npm scripts, instead using webpack / babel
features to allow for rewriting the imports of the lerna packages
so they can be more easily processed as normal local imports. This
would require a number of other changes so a simpler approach is
being taken first to assess the impact.