Remove `step=__` from URL for PipelineRun and TaskRun details pages
Eliminate the need for the `step=__` query param to view the TaskRun
details on both the PipelineRun and TaskRun details pages.

Maintain the default behaviour of selecting the first / first failed
step when loading these pages, but still allow for URL addressability
of the TaskRun details view showing params / status.