Adopt Carbon gray10 theme
Phase 1 of the design update: adopt Carbon gray10 theme with high contrast
gray100 side nav to match the header. This uses Carbon tokens for all colours,
with the exception of 2 instances in the side nav styles which were recommended
by the Carbon team.

The next phase will update the visuals of the TaskTree component in preparation
for changes to how we display conditions and retries and link them to their
owner TaskRuns.

Changes involved:
- use dark style SideNav in preparation for rail adoption
- use dark background for logs and YAML resources
- fix namepaces dropdown theme on Create Secret page
- fix colours and hover state for log buttons
- fix colour and contrast of log trailer
- use Carbon colours for step + task status
- implement workaround for 'light' DataTable on TaskRunDetails and StepDefinition tabs
- update Import Resources to be consistent with other pages