Fix OverflowMenu focus and keyboard nav issue
The latest Carbon update marked the `primaryFocus` prop on OverflowMenuItem
as deprecated, however it's actually not used anymore. This means that when
an OverflowMenu is opened, nothing gets focus by default breaking keyboard
nav.

Use the `selectorPrimaryFocus` prop on the OverflowMenu instead. The default
value does not match anything, so set it to match the first enabled button.