Ensure label / annotation forms fully support keyboard navigation
The 'remove' action on the label and annotation lists in Create Secret,
Create PipelineRun, and Create TaskRun was an icon and therefore
not focusable by keyboard by default. This is both an accessibility
and general usability problem.

Replace the icon with an icon button. This is keyboard navigable
by default, and will be handled correctly by screenreaders and
other assistive technologies.