Update PipelineRun container logic to correctly select TaskRuns with conditions
The `tekton.dev/pipelineTask` label is applied to both the condition check TaskRun
and the owning TaskRun. The previous logic could in some cases incorrectly select
the condition check for display in the UI.

Update the logic to ensure a match for the `tekton.dev/pipelineTask` label is only
counted when there is no accompanying `tekton.dev/conditionCheck` label. This will
ensure the condition is only matched when the selected value appears in the
conditionCheck label, regardless of the order TaskRuns are listed in the data.