Update component stories to use Component Story Format
Component Story Format (CSF) is the recommended format for
stories since StoryBook 5.2.

Update the stories for the components to this new format,
and make proper use of decorators so other Storybook addons
can correctly process the component (e.g. to generate prop tables).

Add @storybook/addon-notes and move some of the longer story
titles to notes for consistency.