Fix dropdown styles on Create screens after latest theme and Carbon updates
The `light` prop on dropdowns needs to be switched to account for the
latest theme update as well as the Carbon 10.16 update.

The standalone pages now have a gray10 background, so the dropdowns
should have a white background to make them clearer.

On the remaining create dialogs the dropdowns should still have a gray
background since the dialog background is white.