Update LogFormat conceal styles so text can still be selected / copied
In environments where the 'conceal' style is supported it is typically
implemented by setting the foreground colour equal to the background
colour. This means that the content can still be viewed by selecting it,
and can also be copy/pasted if needed.

Switch from `visibility: hidden` (which means the content cannot be copied)
to `color: transparent` to match typical implementations.