Update Graph storybook entries to new CSF format
Update the Graph storybook entries to the new Component Story Format
for consistency with all our other stories.

Since the default export is used to provide the name and other config
for display in the storybook UI, this required splitting the single
Graph.stories.js into 1 file per component type.