Update release pipeline to make publishing as latest optional
Publish artifacts to the versioned location first, and optionally
copy all artifacts to 'latest'. The `releaseAsLatest` param controls
both this and tagging the images as 'latest'.

Include the installer manifests in the lockdown.py post-processing
so the oauth-proxy image includes the SHA instead of pointing at
'latest'.

This is more consistent with the approach taken by other Tekton
projects and avoids the need for manual intervention to copy
files to the correct locations after cutting a release.