Remove OpenShift specific install manifests
The main difference between the default Dashboard install and the
OpenShift variant is the inclusion of the oauth-proxy and related
config. Since the oauth-proxy has a number of shortcomings (e.g.
lack of proper session timeout / invalidation) and better
alternatives exist, remove the oauth-proxy config from the
Dashboard install. We have walkthroughs in the Dashboard docs that
describe the steps required to set up oauth2-proxy or Envoy as
examples. Removing oauth-proxy allows us to simplify the release
pipeline and remove the lockdown.py script.

As well as removing the oauth-proxy config, remove the `/v1/endpoint`
and `/v1/ingress` APIs since they're not used by the Dashboard and
were for legacy integration. This allows us to remove our dependency
on the OpenShift client and reduce the amount of code that needs to be
maintained and tested.

Also remove the OpenShift release pipeline since it's incomplete,
requires manual changes, and we should be moving our release process
to the dogfooding cluster similar to other projects anyway. Since this
change removes the lockdown.py script which didn't work on OpenShift
there shouldn't be an issues with running the normal release pipeline
on OpenShift if we wish to do so.