Add tab showing story source in the storybook
Add the `@storybook/storysource` addon which display a tab beside
the existing Knobs and Actions tabs showing the Source code of the
selected story. This provides an easy way for a dev to review how
the functionality demonstrated by the story was achieve.

This will be complemented by the Docs addon in future, providing
more comprehensive documentation of the various props / combinations
supported by each component.