Fix cluster-scoped Kubernetes resource extensions
When a Kubernetes resource extension is created for a cluster-scoped
resource, in some cases this fails to load in the dashboard UI when
the namespace dropdown has selected a value other than 'All Namespaces'.

Update the API layer to check whether the API resource is namespaced
and use this to construct the API request accordingly.

Also update the UI to hide the namespace column for cluster-scoped
resources in the default extension list view `ResourceList`.

Update the API layer to correctly handle extensions displaying
Kubernetes resources from the `core` group, since the API root
is `/api` instead of `/apis/${group}` for these resources.