Ensure FormattedDate always provides a tooltip with the absolute date/time
Update the FormattedDate component to ensure it always provides a tooltip
containing the absolute date/time regardless of whether the main display
value is absolute / relative.

This is import when the value is being displayed in a table with many
columns or other similar cases where wrapping is undesirable and horizontal
space may be constrained. Since the displayed value may be truncated,
the tooltip provides a way to view the full value. This is consistent with
other fields already displayed in tables and elsewhere in the dashboard.