Remove GitHub restriction on Import Resources
The frontend validation on the Import Resources page expected
a `https` GitHub URL for the 'Repository URL' field. This is
unnecessarily restrictive as the PipelineResource type git used
under the covers now supports more than just GitHub.

Switch out the validation to remove the GitHub restriction and
use a well-tested library for parsing the URL to extract the
required fields to use for the PipelineRun labels (name, owner,
host).