Use globally selected namespace by default on Create pages
When opening the Create PipelineResource or Create Secret page,
if the global namespaces dropdown has selected a value other than
All Namespaces, use this as the default selected namespace in
the create page.

If the global namespaces dropdown is set to All Namespaces, do
not select a default namespace in the create page.