Remove unnecessary fetch polyfill from frontend
All supported browsers provide a native `fetch` implementation so
remove `isomorphic-fetch` from the polyfills.

Instead of this, we use `node-fetch` directly in the unit tests
and also update `fetch-mock` to the latest version.

These changes, along with the Node.js 14 update, reduce the execution
time for the full suite of unit tests from ~2.5mins to under 1 minute.