Move Rerun PipelineRun API to frontend
Remove the `/v1/namespaces/<namespace>/rerun` endpoint from the
backend and implement the logic in the frontend instead using the
same backend API used for creating a new PipelineRun. This has a
number of benefits:
- increased consistency in our API usage
- reduces likelihood of version mismatch in the Pipelines client
  breaking the rerun functionality. This has already happened once
- simplifies the backend code and reduces the custom API surface area

This change also avoids an unnecessary 'get' for the PipelineRun
since it's already available in the client in both places where
the rerun functionality is being used.

A similar approach can now be taken for TaskRuns without the need
to introduce a new API in the backend.