Add support for rerunning standalone TaskRuns
Similar to PipelineRuns, add the option to rerun a TaskRun from the
overflow menu on the TaskRuns list, as well as the Rerun button on
the TaskRun details page.

This is only visible / enabled for standalone TaskRuns, i.e. TaskRuns
that are not owned by a PipelineRun.

As part of this, the Rerun button component was refactored to remove
pipeline-specific terminology and to allow for reuse.