Fix for LogFormat component handling consecutive carriage returns
If log content contains multiple consecutive carriage returns ('\r')
this causes issues with the ANSI parser resulting in an error
displayed to the user and no logs output.

Specifically, the use of '.' and the 'm' (multiline) flag in the
regex for the ANSI parser fails in this case. Instead, it's
recommended to use '[^]' instead of '.' when operating in multiline
mode. This ensures we can match these and other escape sequences
correctly across lines. ES2018 introduces the 's' (dotAll) flag for
similar purpose, but this is not yet supported in all browsers.