Fix update handlers in extensions and ClusterTasks reducers
Tests for the 'UPDATED' case in the extensions and ClusterTasks
reducers were starting with an empty state rather than having
an existing resource to ensure the update was being handled
correctly. As a result we missed the fact that the updates were
not being applied as expected, instead in most cases any changes
would immediately be overwritten by the previous state.

This lead to inconsistent updates of the ClusterTask details
page as well as extensions in some cases.

Fix the tests to ensure an existing resource is present in the
state before testing the update (verified this failed on previous
implementation) and fix the reducers to ensure the update resource
is applied on top of the existing state.