Fix loading of CRD extensions when a specific namespace is selected
The side nav should always display extensions, regardless of the
globally selected namespace. The resulting UI produced by the extension
will be subject to the filter, but its presence/absence in the nav
should not depend on this.

Since we already check for the single namespace mode, we can safely
change the fallback to query all namespaces (i.e. use the non-namespaced
API call) instead of using the globally selected namespace.