Update README to point at 0.10.2 docs
Update the table of doc versions to include 0.10.2 and update
the supported Triggers version.

Move older Dashboard versions to a collapsed section below the
main table, keeping only the latest patch version for each minor
release in the main table.

Also update the section title to match the expected link from
the docs on the website.