Refactor TaskRun / PipelineRun component trees to separate definitions from runtime data
Update the data processing done by the TaskRun/PipelineRun containers and their component
trees to separate the Task definitions from the TaskRun runtime data instead of using
our many custom representations that combined the two in different ways depending on need.

This should help to resolve confusion around the origin of a number of fields used in the
code, make it easier to maintain in future / easier for new contributors to understand,
and also allows for additional functionality that was cumbersome or essentially impossible
to achieve before.

Since we're no longer bound by the definitions to construct the bulk of the UI, we can
now more easily display generated steps for PipelineResources etc. giving a more complete
representation of the run.

This change has a dependency on Pipelines 0.17 for the correct step order, but otherwise
should continue to work with earlier versions.