Fix for websocket connection blocked by CSP in Safari
Safari has not correctly implemented the `connect-src` directive
according to the (updated) Content Security Policy specification.

This means that `connect-src: 'self'` does not include websocket
connections to the same host.

See https://github.com/w3c/webappsec-csp/issues/7 and
https://bugs.webkit.org/show_bug.cgi?id=201591 for details.

Explicitly allow `ws:` and `wss:` in `connect-src` to work around
this for now. Hopefully we can revisit this and remove them when
Safari updates their implementation to follow the spec for this case.