Fix some Carbon themeing issues
Carbon colour tokens had the wrong values in some of the components
due to the way we were importing the variables.

Ensure we're setting the theme correctly when importing variables
to ensure the colour tokens have the right values. Also update
the Storybook to set the expected background colour for each
component so developers can get a better preview of what they'll
actually look like in use.

Also add `stylelint` to help with tracking places in the styles
where we're not currently using Carbon tokens.