Remove duplicate test utils
Remove the duplicate test utils from the `src/` folder, instead
using the ones from the components package.

Also update the utils to use the `react-testing-library` `wrapper`
field instead of directly wrapping the provided component as the
primary parameter to the `render` function. This is the recommended
approach and also allows us to resolve a number of warnings in the
test log output relating to modifying `<Router history>`