Remove publish-images.sh
`publish-images.sh` was only used by the prow postsubmit job for
the Dashboard repo. It is obsolete, we've changed the build
and release process considerably since this was originally written,
and now have nightly releases running on dogfooding which produce
the full expected set of release artifacts.

The postsubmit job has been removed in
https://github.com/tektoncd/plumbing/pull/648 so it's safe to remove
the script too.