Move global namespaces dropdown to list page header
Move the global namespaces dropdown from the side nav to the
header of the list pages.

To avoid duplication and to help ensure consistency, introduce
a new ListPageLayout container which is responsible for providing
the common layout and functionality for the list pages.

Moving the namespaces dropdown allows us to resolve a number of
long-standing issues:
- remove code duplication required to redirect to list pages
  when changing namespace from details page as this scenario
  is no longer possible
- a namespaced view will now always contain the namespace
  in the URL, providing consistency as it no longer behaves
  differently when navigating from a namespaced / non-namespaced
  view to a namespaced list. This also means that URLs are
  reliable and can be shared to produce the same view the user
  is currently viewing

It also allows us to remove a large amount of custom styling
in SideNav.scss which was required to style the dropdown to
match the dark theme used in the nav.

One final change that was required to avoid potentially
expensive API calls to build the nav for extensions is the
addition of an optional `namespaced` field (defaults to `true`)
on the kubernetes-resource extensions. This is used to ensure
we generate the correct URL in the side nav (i.e. namespaced /
non-namespaced) regardless of the current URL / globally
selected namespace.