Update to webpack 5
Update webpack to v5. This also requires updating all of the
supporting loaders / plugins since there were a number of breaking
changes in events and configuration.

This produces minimal reduction in chunk sizes, which is expected,
as I've kept the configuration as close as possible to what we were
using with webpack 4 so we can identify any new issues before making
further changes to optimise our config.