Display placeholders for pipeline tasks that haven't begun executing yet
On the PipelineRun details page, display placeholders for pipeline tasks
that don't yet have a corresponding TaskRun. Fill in spec details where
available so the user can still view the definition.

Once the pipeline task begins executing the placeholder will be replaced
with the real TaskRun and corresponding details.

In the case of Conditions, we make a best effort to display placeholders
for the TaskRuns generated to evaluate them. Once they execute we will
display the correct TaskRun details, however before this is available
the placeholder may be incomplete and may become unselected as the
PipelineRun updates. This is due to a lack of required runtime information
to produce a consistent link between the resources. Since Conditions are
deprecated, we're not planning to invest further effort in improving this.

This update also changes the ordering of TaskRun in the TaskTree component.
Instead of sorting by start time of the first step in each TaskRun, they're
now sorted in order of definition in the Pipeline.

A nice (subj.) side effect of this is that retries for a given TaskRun
will be displayed immediately following the corresponding TaskRun rather
than interspersed among other TaskRuns in an often unpredictable order.
Future changes to the display of retries will further improve this grouping.