Add ability to filter runs by status
Add filter dropdown on the TaskRuns and PipelineRuns lists
to allow the user to filter the displayed runs by status.

Add a new StatusFilterDropdown which encapsulates the display /
translations of the status options, while leaving the containers
responsible for performing the actual filtering of the data they
provide to the list UI components.