Update empty state strings for consistency and accuracy
Update the strings displayed on empty tables to be more consistent
across resource types and also to be more accurate.

Since we've added label filtering support the 'no resources found'
message has been a bit misleading, it should really be 'no matching
resources found'.

This is even more obvious with the recent addition of status filtering
on the TaskRuns and PipelineRuns lists.

Update the strings used for all namespaces, cluster resources, and
selected namespace empty states.