Add ability to delete Pipelines, Tasks, and ClusterTasks
Add the ability to delete Pipelines, Tasks, and ClusterTasks, both
individually and in bulk via the Carbon Table batch actions toolbar.

There's one slight difference in the user experience for these resource
types compared to the existing delete functionality for PipelineRuns
and TaskRuns. That is, the delete action for a single resource is
exposed as a button inline on the table row rather than being nested
in an overflow menu. This is because we only have at most 2 actions
on a row, one being the link to the details page for the resource.
Carbon guidelines recommend keeping the actions inline when there are
fewer than 3 actions.