Reduce code duplication for delete modals and disambiguate resources in different namespaces
Add a new DeleteModal component to handle the common layout / styling
for all of our views that use it. Update all containers that provide
delete functionality to use the new component.

The new modal uses a Table instead of UnorderedList and will display
the namespace of the selected resources (when in 'All Namepsaces' mode
and viewing namespaced resources)

As part of this change I've also updated the Secrets container to be
more consistent with the use of complete resources (referenced by id)
we have for other resource types and to avoid string manipulation to
pass name/namespace around. This also removes the need for the
SecretsDeleteModal as it's now using the common pattern and can use
the generic DeleteModal.