Update installer script to build multi-arch image only on release action
When a user is installing the Dashboard in a dev environment using the
installer script there's no need to build for other architectures.

Update the script to pass the `--platform=all` arg to `ko` only for
the release action and update the e2e test script to ensure we're
actually exercising the release action and not just the build.