Handle properties fallback when running in client mode
When the Dashboard frontend is running in client mode (i.e.
`kubectl proxy --www...`), set the fallback values for the
install properties (e.g. ReadOnly, IsTriggersInstalled) so
the Dashboard is more useful in this case.

Future updates will add more fine-grained access checks so
we only display options that are actually available depending
on the user's permissions.