Update Dashboard screenshot on main readme
Update the screenshot displayed on the main readme at the project root
to reflect the latest design updates for the nav and the TaskTree.

Switch image formats from PNG to JPG for >80% file size reduction
with little noticeable impact on image quality. We've made a similar
switch for the Katacoda tutorials on the website recently.