Fix Storybook
Storybook has been broken for a while (?) due to incompatability
between the Storybook version we were using and our webpack setup,
in particular the version of the html-webpack-plugin.

Update all storybook packages and the html-webpack-plugin to the
latest available versions.

With this update storybook builds (both dev + prod) are working
as expected.