Remove dependency on Tekton go clients
Remove the dependency on Tekton Pipeline and Tekton Triggers
go clients in the Dashboard backend.

Use the kubernetes dynamic client (and informers) since we're not
inspecting or manipulating the resources in any way that would benefit
from having typed objects. Removing the dependency on the typed clients
means we're less tightly coupled to the Pipelines and Triggers releases.