Simplify integration tests to focus on Dashboard functionality
The integration tests were previously executing a PipelineRun
which cloned a repo, built an image, and deployed an app to
the cluster. These are not Dashboard features and so are not
particularly useful in verifying changes are not introducing
bugs in the Dashboard itself.

Simplify the test Pipeline to output a simple Hello World
message so we can focus on the Dashboard APIs and not have to
worry about additional deployments, registries, etc. Ensure
we're actually testing the Dashboard API in read-only mode
too.

This removes the dependency on `helm`, updates `yq` to the
latest release, and updates `newman` to the latest release.

Update test resources to tekton.dev/v1beta1

Simplify the test logic to verify the TaskRun produces the
expected output.

Fix local-e2e-tests script to work with `kind` without an
additional docker registry.