display language according to the language setting
We changed the language displayed on the screen according to the language setting of the browser.
Valid languages ​​are defined in locales.supported of config_frontend / config.json.
If not supported, the default language (English) will be displayed.

We have added Japanese to the locales.supported parameter in config_frontend / config.json.