Update `build:analyze` script to provide more accurate results
Update the `build:analyze` npm script to pass the bundleDir param
so the results presented by the tool are more representative of the
expected output after tree-shaking.

Also update editorconfig to enforce trailing newline in package.json
as added by `npm` to avoid unwated noise in diffs when manually
editing the file.