Extend e2e test script to run tests on non-Intel architectures
At this moment integrationtests can be executed with default e2e-tests.sh
script only for amd64 architecture, because image during setup process is
built for amd64 by default. Addition of the PLATFORM env variable to e2e
test script allows to build dashboard image for needed architecture via
--platform parameter in ko tool.
In case if PLATFORM variable is not specified, there will be current behaviour.

Also SKIP_BUILD_TEST env varaible is added to skip dashboard build verification
as it is done for each nightly build and is not required for testing on
non-Intel architecture specifically.

Signed-off-by: Yulia Gaponenko <yulia.gaponenko1@de.ibm.com>