Simplify version checks in back end
Simplify the back end code that's used to get the version of the
installed Tekton projects in the cluster for display on the About
page. Also remove the `IsOpenShift` flag as it no longer provides
any value.

All supported versions have standardised on the set of labels
present on deployments, including both the project name and version.

This is common to manual and operator installs. Tested with the
following operator installs:
- OpenShift Pipelines Operator on CRC 1.24 (OpenShift 4.7.2)
- Tekton Operator v0.22.0-1 on kind (kubernetes 1.20.2)