Add new containers for the Trigger and ClusterInterceptor pages
Add new list and details pages for the Trigger and ClusterInterceptor
CRDs.

Reuse the `Trigger` component already used by the EventListener details
page to display details of the new `Trigger` resource type. Update the
component to handle both the `Trigger` resource, and the existing
EventListener triggers definitions.

Update the EventListener details to link to the new resources when
refs to them are found.