Add custom hook to refetch data when websocket reconnects
Add a custom React hook that can be used to perform an action
when the websocket is reconnected (but not on initial connection).
This is useful in our containers so we can avoid refetching when
the websocket disconnects (default behaviour if we use `webSocketConnected`
as a dependency for `useEffect`).

This avoid unnecessary and unwanted requests while keeping the container
logic relatively simple. In future with `react-query` or similar libraries
we can easily change the central `useWebSocketReconnected` hook as needed
without having to update each container.

Continue refactoring containers to be functional components, using
the new hook as appropriate.