Fix runtime error after delete cluster task
The error happens when user click delete button in confirm delete modal.
The reason is when open the modal it set the callback directly to
`cancelSelection` via `setCancelSelection` but `React.useState` treat
that it must be return a new state value when passing a function to it.
To fix this issue, passing a function that return a callback that want
to set instead.