Remove the ClusterRole aggregation from the default manifests
ClusterRole aggregation is only used to support Dashboard extensions
and requires additional permissions on the cluster. This causes
problems in many scenarios and currently affects all users even if
they're not using extensions.

Move the aggregation rule behind an installer flag so users who
need it can still easily generate the required config but other
users benefit from the reduced permissions required for a default
install.

Document the flag in the extensions doc and describe the requirements
for achieving the same result manually for users not using the
installer script.

Also update the installer docs and help command to describe the
new option.