Update SideNav to hide when collapsed
Hide the SideNav when collapsed instead of displaying the rail on
the left of the window. The rail has a number of issues such as
inconsistent positioning of icons, unexpected behaviour when
NavMenus are expanded/collapsed, over-eager response to hover events,
and generally causes frustration when users are attempting to use it.

Update unit tests to account for the new behaviour while also
removing unnecessary use of the redux store in some tests.