Add hook to keep page title in sync with displayed resource(s)
Add a custom hook which will keep the `document.title` in sync
with the display resource(s) as the user navigates through the
application. On a given resource details page, respond to changes
in the resource name. Responding to changes caused by navigation
to another resource kind is already handled via the router as a
new page container will be mounted and its `useTitleSync` hook
will take over.