Update DataTableSkeleton to support `size` prop
Update DataTableSkeleton to support a `size` prop with the same
values supported by the Carbon `Table` component, and update our
`Table` wrapper component to pass the value through.

Add control for the supported size values to the storybook.

This will be used on the updated About page loading state as part
of the react-query migration.