Add initial react-query utilities
Add a custom hook for fetching a list of namespaced resources
similar to our existing `fetchNamespacedCollection` action creator.
The difference with this is that we won't need action creators,
selectors, or the reducer backing it, all of the caching functionality
is handled internally by react-query while giving us control over
config (e.g. cache time, stale time, retries) if desired.

Add a custom hook to handle websocket events, updating the react-query
cache directly for any matching queries. This is a replacement for
the existing websocket middleware and events handlers we've added with
redux.

These are not yet wired into the application, this will be handled
in another PR.