Move more resources to react-query
Move the conditions, pipelineResources, and serviceAccounts
slices of the state from redux to react-query.

Remove the tests for the reducerCreator that relied on the
pipelineResources implementation as this set of utilities
will be removed in one of the next PRs anyway so there's not
much point in updating those tests at this point.