Update style of links in tables to match Carbon guidelines
Adopt the Carbon `Link` component so we correctly inherit the styles
expected for Carbon. This includes removing the default underline
from links in tables since this is reserved for inline links only.

This also brings some minor changes to the hover, focus, and active
states, mostly the text colour, and uses colour tokens so links are
still legible in dark mode.