Add auto-scroll to logs
Logs of step runs can get very big. While the step run has not terminated, the p
age will keep scrolling down the page as needed. This behaviour can be stopped once the user scrolls up so that the bottom of the log is no longer visible. This behaviour can be resumed once the user scrolls back down to the log bottom.

This feature is activated by a boolean flag (which is turned to true in the dashboard ui) not to affect the users of the reusable component `<Log` in edge cases (the component followed by other elements in the page,component inside a non-window scrollable parent.....)

Signed-off-by: Mostafa-Abdelwahab <mostafa.abdelwahab@ibm.com>