Add scroll buttons to logs
Step-runs with huge logs require the user to manually scroll all the way down the page. A click/press-to-scroll-to-page-top button will now appear when the top of the logs is above the viewport and a click/press-to-scroll-to-page-bottom button will now appear when the bottom of the logs is below the viewport.

The button presence is updated at each Log component update or scroll. For a logically-complete solution, we would need to use a ResizeObserver. This feature is behing a flag to not affect the users of the reusable `<Log` component with edge cases (a non-window-scrollable parent).

Signed-off-by: Mostafa-Abdelwahab <mostafa.abdelwahab@ibm.com>