Add support for overriding the default log polling cut-off
In some cases consumers of the PipelineRun component may want to
extend the log polling beyond the default cut-off point (we currently
make 1 final request after the step is terminated).

Provide a new prop on the PipelineRun component which is passed
through to the Log component allowing consumers to override
the default behaviour and force polling to continue until some
externally determined status is met.