Update UI components to support displaying TaskRun Pod details
Add a new tab to the TaskRunDetails to display the associated Pod
and any relevant events. This tab is only present when the required
Pod information is provided to the component.

This will not be exposed in the Dashboard application until a
future change fetches the Pod information in both the PipelineRun
and TaskRun containers and passes it to the UI components.