Update proxy to support websockets
Our current proxy does not support connection upgrades, which means we do not
have a clean way to access the built-in support for websockets provided by the
Kubernetes API server. Instead, the Dashboard provides a single websocket
connection which provides updates for all supported resources.

This has a number of drawbacks such as:
- not being able to selectively register to receive updates for:
  - resources in a specific namespace
  - a particular kind of resource
  - a single specific resource
- not supporting real time updates for resource-based extensions

This means that all pages in the Dashboard incur an unnecessary overhead by
processing updates for resources they may never require.

Add a new proxy implementation which handles the connection upgrades, while
also providing the ability to support more fine-grained permissions and
user-based auth which we want to add in future.

Migrate the existing endpoints to the new server/proxy setup, including the
existing websocket connection as it's still in use but will be replaced in the
near future.

Remove the service-based extension support as the webhooks-extension was
deprecated after v0.7.0 and has since been deleted. Similar functionality could
be added via a small modification to the resource-based extensions and using
the new proxy to access the services.